backend/requirements.txt
# Core Dependencies
google-generativeai>=0.3.0  # Google AI Studio SDK
pinecone-client>=2.2.0
kafka-python>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0

# Data Processing (optional - not required for demo)
# pandas>=2.1.0

# Logging
loguru>=0.7.0

# CORS
fastapi-cors>=0.0.6

# Optional: For advanced features
# schedule==1.2.0  # For scheduled tasks
# asyncio==3.4.3  # Async support